[dependency] First set of adjustments to PR #189
