+++ asistente_ladm_col/gui/qgis_model_baker/dlg_import_data.py
@@ -19,8 +19,8 @@
 import os
 import re
 
-from QgisModelBaker.libili2db import iliexporter
-from QgisModelBaker.libili2db.ili2dbconfig import (ExportConfiguration,
+from QgisModelBaker.libili2db import iliimporter
+from QgisModelBaker.libili2db.ili2dbconfig import (ImportDataConfiguration,
                                                    BaseConfiguration)
 from QgisModelBaker.libili2db.ili2dbutils import color_log_text
 from QgisModelBaker.libili2db.ilicache import IliCache
@@ -35,28 +35,25 @@ from qgis.PyQt.QtGui import (QColor,
 from qgis.PyQt.QtWidgets import (QDialog,
                                  QSizePolicy,
                                  QListWidgetItem,
-                                 QMessageBox,
                                  QDialogButtonBox)
 from qgis.core import Qgis
 from qgis.gui import QgsGui
 from qgis.gui import QgsMessageBar
 
-from ...config.general_config import DEFAULT_HIDDEN_MODELS
-from ...lib.dbconnector.pg_connector import PGConnector
+from ...config.general_config import (DEFAULT_EPSG,
+                                      DEFAULT_INHERITANCE,
+                                      DEFAULT_HIDDEN_MODELS)
 from ...utils import get_ui_class
 from ...utils.qt_utils import (Validators,
                                FileValidator,
-                               make_save_file_selector,
                                make_file_selector,
+                               make_save_file_selector,
                                OverrideCursor)
 from ...resources_rc import *
 
-DIALOG_UI = get_ui_class('model_baker/dlg_export_data.ui')
+DIALOG_UI = get_ui_class('qgis_model_baker/dlg_import_data.ui')
 
-class DialogExportData(QDialog, DIALOG_UI):
-    ValidExtensions = ['xtf', 'itf', 'gml', 'xml']
-    current_row_schema = 0
-    
+class DialogImportData(QDialog, DIALOG_UI):
     def __init__(self, iface, db, qgis_utils):
         QDialog.__init__(self)
         self.setupUi(self)
@@ -69,23 +66,20 @@ class DialogExportData(QDialog, DIALOG_UI):
         self.ilicache.refresh()
 
         self.type_combo_box.clear()
-        self.type_combo_box.addItem(QCoreApplication.translate('DialogExportData','PostgreSQL/PostGIS'), 'ili2pg')
-        self.type_combo_box.addItem(QCoreApplication.translate('DialogExportData','GeoPackage'), 'ili2gpkg')
+        self.type_combo_box.addItem(QCoreApplication.translate('DialogImportData','PostgreSQL/PostGIS'), 'ili2pg')
+        self.type_combo_box.addItem(QCoreApplication.translate('DialogImportData','GeoPackage)'), 'ili2gpkg')
         self.type_combo_box.currentIndexChanged.connect(self.type_changed)
         self.type_changed()
 
         self.xtf_file_browse_button.clicked.connect(
-            make_save_file_selector(self.xtf_file_line_edit, title=QCoreApplication.translate('DialogExportData','Save in XTF Transfer File'),
-                                    file_filter=QCoreApplication.translate('DialogExportData','XTF Transfer File (*.xtf);;Interlis 1 Transfer File (*.itf);;XML (*.xml);;GML (*.gml)'), extension='.xtf', extensions=['.' + ext for ext in self.ValidExtensions]))
-        self.xtf_file_browse_button.clicked.connect(self.xtf_browser_opened_to_true)
-        self.xtf_browser_was_opened = False
-        
+            make_file_selector(self.xtf_file_line_edit, title=QCoreApplication.translate('DialogImportData','Open Transfer or Catalog File'),
+                               file_filter=QCoreApplication.translate('DialogImportData','Transfer File (*.xtf *.itf);;Catalogue File (*.xml *.xls *.xlsx)')))
+
         self.validators = Validators()
-        fileValidator = FileValidator(pattern=['*.' + ext for ext in self.ValidExtensions], allow_non_existing=True)
-        self.xtf_file_line_edit.setPlaceholderText(QCoreApplication.translate('DialogExportData', "[Name of the XTF to be created]"))
+        self.xtf_file_line_edit.setPlaceholderText(QCoreApplication.translate('DialogImportData', "[Name of the XTF to be created]"))
+        fileValidator = FileValidator(pattern=['*.xtf', '*.itf', '*.xml'])
         self.xtf_file_line_edit.setValidator(fileValidator)
-        self.xtf_file_line_edit.textChanged.connect(self.validators.validate_line_edits)
-        self.xtf_file_line_edit.textChanged.connect(self.xtf_browser_opened_to_false)
+        self.xtf_file_line_edit.textChanged.connect(self.update_import_models)
         self.xtf_file_line_edit.textChanged.emit(self.xtf_file_line_edit.text())
 
         # PG
@@ -93,19 +87,25 @@ class DialogExportData(QDialog, DIALOG_UI):
         self.db_connect_label.setText(self.db.dbname)
         self.connection_setting_button.clicked.connect(self.show_settings)
 
-        self.connection_setting_button.setText(QCoreApplication.translate('DialogExportData', 'Connection Settings'))
+        self.connection_setting_button.setText(QCoreApplication.translate('DialogImportData', 'Connection Settings'))
 
         # GPKG
-        self.gpkg_file_line_edit.setPlaceholderText(QCoreApplication.translate('DialogExportData', "[Name of the Geopackage to be created]"))
-        self.gpkg_file_browse_button.clicked.connect(make_file_selector(self.gpkg_file_line_edit, title=QCoreApplication.translate('DialogExportData','Open GeoPackage database file'),file_filter=QCoreApplication.translate('DialogExportData','GeoPackage Database (*.gpkg)')))
-        gpkgFileValidator = FileValidator(pattern='*.gpkg')
+        self.gpkg_file_line_edit.setPlaceholderText(QCoreApplication.translate('DialogImportData', "[Name of the Geopackage to be created]"))
+
+        self.gpkg_file_browse_button.clicked.connect(
+            make_save_file_selector(self.gpkg_file_line_edit, title=QCoreApplication.translate('DialogImportData','Save in GeoPackage database file'),
+                                    file_filter=QCoreApplication.translate('DialogImportData','GeoPackage Database (*.gpkg)'), extension='.gpkg'))
+
+        gpkgFileValidator = FileValidator(pattern='*.gpkg', allow_non_existing=True)
         self.gpkg_file_line_edit.setValidator(gpkgFileValidator)
+
+        self.gpkgSaveFileValidator = FileValidator(pattern='*.gpkg', allow_non_existing=True)
+        self.gpkgOpenFileValidator = FileValidator(pattern='*.gpkg')
         self.gpkg_file_line_edit.textChanged.connect(self.validators.validate_line_edits)
         self.gpkg_file_line_edit.textChanged.emit(self.gpkg_file_line_edit.text())
 
         # LOG
-        self.log_config.setTitle(QCoreApplication.translate('DialogExportData', 'Show log'))
-        self.log_config.setFlat(True)
+        self.log_config.setTitle(QCoreApplication.translate('DialogImportData', 'Show log'))
 
         self.bar = QgsMessageBar()
         self.bar.setSizePolicy(QSizePolicy.Minimum, QSizePolicy.Fixed)
@@ -115,19 +115,73 @@ class DialogExportData(QDialog, DIALOG_UI):
         self.buttonBox.accepted.connect(self.accepted)
         self.buttonBox.clear()
         self.buttonBox.addButton(QDialogButtonBox.Cancel)
-        self.buttonBox.addButton(QCoreApplication.translate('DialogExportData', 'Export data'), QDialogButtonBox.AcceptRole)
+        self.buttonBox.addButton(QCoreApplication.translate('DialogImportData', 'Import data'), QDialogButtonBox.AcceptRole)
         self.buttonBox.addButton(QDialogButtonBox.Help)
         self.buttonBox.helpRequested.connect(self.show_help)
 
     def showEvent(self, event):
-        # update after create dialog
         self.update_schema_names_model()
-        self.update_models_names(self.get_checked_schema())
         self.restore_configuration()
+
+    def update_import_models(self):
+
+        message_error = None
+
+        if not self.xtf_file_line_edit.text().strip():
+            color = '#ffd356'  # Light orange
+            self.import_models_qmodel = QStandardItemModel()
+            self.import_models_list_view.setModel(self.import_models_qmodel)
+        else:
+
+            if os.path.isfile(self.xtf_file_line_edit.text().strip()):
+                color = '#fff'  # White
+
+                self.import_models_qmodel = QStandardItemModel()
+                models_name = self.find_models_xtf(self.xtf_file_line_edit.text().strip())
+
+                for model_name in models_name:
+                    if not model_name in DEFAULT_HIDDEN_MODELS:
+                        item = QStandardItem(model_name)
+                        item.setCheckable(False)
+                        item.setEditable(False)
+                        self.import_models_qmodel.appendRow(item)
+
+                if self.import_models_qmodel.rowCount() > 0:
+                    self.import_models_list_view.setModel(self.import_models_qmodel)
+                else:
+                    message_error = QCoreApplication.translate('DialogImportData',
+                                                               'The XTF file does not register models, please verify')
+                    color = '#ffd356'  # Light orange
+                    self.import_models_qmodel = QStandardItemModel()
+                    self.import_models_list_view.setModel(self.import_models_qmodel)
+            else:
+                message_error = QCoreApplication.translate('DialogImportData', 'Please set a valid XTF file')
+                color = '#ffd356'  # Light orange
+                self.import_models_qmodel = QStandardItemModel()
+                self.import_models_list_view.setModel(self.import_models_qmodel)
+        self.xtf_file_line_edit.setStyleSheet('QLineEdit {{ background-color: {} }}'.format(color))
+
+        if message_error:
+            self.txtStdout.setText(message_error)
+            self.show_message(message_error, Qgis.Warning)
+            self.import_models_list_view.setFocus()
+            return
+
+    def find_models_xtf(self, xtf_path):
+        models_name = list()
+        pattern = re.compile(r'<MODEL[^>]*>(?P<text>[^<]*)</MODEL>')
+        with open(xtf_path, 'r') as f:
+            for txt in pattern.finditer(f.read()):
+                model_tag = str(txt.group(0))
+                name = re.findall( 'NAME="(.*?)"', model_tag, re.IGNORECASE)
+                models_name.extend(name)
+        return models_name
+
 
     def update_schema_names_model(self):
         res, msg = self.db.test_connection()
         schema_names = self.db._schema_names_list()
+        self.schema_names_list_widget.clear()
 
         if schema_names:
             for schema_name in schema_names:
@@ -139,17 +193,18 @@ class DialogExportData(QDialog, DIALOG_UI):
             default_item = self.schema_names_list_widget.item(0)
             default_item.setCheckState(Qt.Checked)
         else:
-            self.schema_names_list_widget.clear()
-            self.export_models_list_view.setModel(QStandardItemModel())
-            message_error = 'There are no schemes to export into the database. Please select another database.'
-            self.txtStdout.setText(QCoreApplication.translate('DialogExportData', message_error))
+            message_error = 'There are no schemes to import into the database. Please select another database.'
+            self.txtStdout.setText(QCoreApplication.translate('DialogImportData', message_error))
             self.show_message(message_error, Qgis.Warning)
+            # try:
+            #     self.show_message(message_error, Qgis.Warning)
+            # except AttributeError:
+            #     pass
 
         self.schema_names_list_widget.currentRowChanged.connect(self.on_current_row_changed_schema_names)
         self.schema_names_list_widget.itemChanged.connect(self.on_itemchanged_schema_name)
 
     def on_itemchanged_schema_name(self, selected_item):
-
         # disconnect signal to do changes in the items
         self.schema_names_list_widget.itemChanged.disconnect(self.on_itemchanged_schema_name)
 
@@ -164,46 +219,18 @@ class DialogExportData(QDialog, DIALOG_UI):
         item.setCheckState(Qt.Checked)
         item.setSelected(True)
 
-        # Update list view with models name info
-        schema_name = item.text()
-        self.update_models_names(schema_name)
-
         # connect signal to check when the items change
         self.schema_names_list_widget.itemChanged.connect(self.on_itemchanged_schema_name)
 
 
     def on_current_row_changed_schema_names(self, current_row):
-
         for index in range(self.schema_names_list_widget.count()):
             item = self.schema_names_list_widget.item(index)
             item.setCheckState(Qt.Unchecked)
 
         item = self.schema_names_list_widget.item(current_row)
-
         if item:
             item.setCheckState(Qt.Checked)
-            # Update list view with models name info
-            schema_name = item.text()
-            self.update_models_names(schema_name)
-
-    def update_models_names(self, dbschema):
-        _db_connector = PGConnector(self.db.get_uri_without_schema(), dbschema)
-        res, msg = _db_connector.test_connection()
-        self.export_models_qmodel = QStandardItemModel()
-        if res:
-            db_models = _db_connector._get_models(dbschema)
-            if db_models:
-                for db_model in db_models:
-                    regex = re.compile(r'(?:\{[^\}]*\}|\s)')
-                    for modelname in regex.split(db_model['modelname']):
-                        if modelname and modelname not in DEFAULT_HIDDEN_MODELS:
-                            item = QStandardItem(modelname.strip())
-                            item.setCheckable(False)
-                            item.setEditable(False)
-                            self.export_models_qmodel.appendRow(item)
-
-        self.export_models_list_view.setModel(self.export_models_qmodel)
-
 
     def get_checked_schema(self):
         checked_schema = None
@@ -216,8 +243,8 @@ class DialogExportData(QDialog, DIALOG_UI):
 
     def get_ili_models(self):
         ili_models = list()
-        for index in range(self.export_models_qmodel.rowCount()):
-            item = self.export_models_qmodel.item(index)
+        for index in range(self.import_models_qmodel.rowCount()):
+            item = self.import_models_qmodel.item(index)
             ili_models.append(item.text())
         return ili_models
 
@@ -232,45 +259,34 @@ class DialogExportData(QDialog, DIALOG_UI):
 
         configuration = self.updated_configuration()
 
+        if not os.path.isfile(self.xtf_file_line_edit.text().strip()):
+            message_error = 'Please set a valid XTF file before importing data. XTF file does not exist'
+            self.txtStdout.setText(QCoreApplication.translate('DialogImportData', message_error))
+            self.show_message(message_error, Qgis.Warning)
+            self.xtf_file_line_edit.setFocus()
+            return
+
+
         if not self.xtf_file_line_edit.validator().validate(configuration.xtffile, 0)[0] == QValidator.Acceptable:
-            message_error = 'Please set a valid XTF file before exporting data.'
-            self.txtStdout.setText(QCoreApplication.translate('DialogExportData',message_error))
+            message_error = 'Please set a valid XTF before importing data.'
+            self.txtStdout.setText(QCoreApplication.translate('DialogImportData',message_error))
             self.show_message(message_error, Qgis.Warning)
             self.xtf_file_line_edit.setFocus()
             return
 
         if not self.get_ili_models():
-            message_error = QCoreApplication.translate('DialogExportData','Please set a valid Schema to export. This schema no have information to export')
+            message_error = QCoreApplication.translate('DialogImportData','The selected XTF file does not have information according to the LADM-COL model to import.')
             self.txtStdout.setText(message_error)
             self.show_message(message_error, Qgis.Warning)
-            self.export_models_list_view.setFocus()
-            return
-        
-        if not configuration.iliexportmodels:
-            message_error = QCoreApplication.translate('DialogExportData','Please set a model before exporting data.')
-            self.txtStdout.setText(message_error)
-            self.show_message(message_error, Qgis.Warning)
-            self.export_models_list_view.setFocus()
+            self.import_models_list_view.setFocus()
             return
 
         if self.type_combo_box.currentData() == 'ili2gpkg':
             if not configuration.dbfile or self.gpkg_file_line_edit.validator().validate(configuration.dbfile, 0)[0] != QValidator.Acceptable:
-                message_error = QCoreApplication.translate('DialogExportData', 'Please set an existing database file before creating the project.')
+                message_error = QCoreApplication.translate("DialogImportData", 'Please set a valid database file before creating the project.')
                 self.txtStdout.setText(message_error)
                 self.show_message(message_error, Qgis.Warning)
                 self.gpkg_file_line_edit.setFocus()
-                return
-
-        # If xtf browser was opened and the file exists, the user already chose
-        # to overwrite the file
-        if os.path.isfile(self.xtf_file_line_edit.text().strip()) and not self.xtf_browser_was_opened:
-            self.msg = QMessageBox()
-            self.msg.setIcon(QMessageBox.Warning)
-            self.msg.setText(QCoreApplication.translate('DialogExportData',"{filename} already exists.\nDo you want to replace it?").format(filename=os.path.basename(self.xtf_file_line_edit.text().strip())))
-            self.msg.setWindowTitle(QCoreApplication.translate('DialogExportData',"Save in XTF Transfer File"))
-            self.msg.setStandardButtons(QMessageBox.Yes | QMessageBox.No)
-            msg_box = self.msg.exec_()
-            if msg_box == QMessageBox.No:
                 return
             
         with OverrideCursor(Qt.WaitCursor):
@@ -281,30 +297,30 @@ class DialogExportData(QDialog, DIALOG_UI):
             self.txtStdout.setTextColor(QColor('#000000'))
             self.txtStdout.clear()
 
-            exporter = iliexporter.Exporter()
+            dataImporter = iliimporter.Importer(dataImport=True)
 
-            exporter.tool_name = self.type_combo_box.currentData()
-            exporter.configuration = configuration
+            dataImporter.tool_name = self.type_combo_box.currentData()
+            dataImporter.configuration = configuration
 
             self.save_configuration(configuration)
 
-            exporter.stdout.connect(self.print_info)
-            exporter.stderr.connect(self.on_stderr)
-            exporter.process_started.connect(self.on_process_started)
-            exporter.process_finished.connect(self.on_process_finished)
+            dataImporter.stdout.connect(self.print_info)
+            dataImporter.stderr.connect(self.on_stderr)
+            dataImporter.process_started.connect(self.on_process_started)
+            dataImporter.process_finished.connect(self.on_process_finished)
 
             self.progress_bar.setValue(25)
 
             try:
-                if exporter.run() != iliexporter.Exporter.SUCCESS:
+                if dataImporter.run() != iliimporter.Importer.SUCCESS:
                     self.enable()
                     self.progress_bar.hide()
-                    self.show_message(QCoreApplication.translate('DialogExportData', 'An error occurred when exporting the data. For more information see the log...'), Qgis.Warning)
+                    self.show_message(QCoreApplication.translate('DialogImportData', 'An error occurred when importing the data. For more information see the log...'), Qgis.Warning)
                     return
             except JavaNotFoundError:
                 self.txtStdout.setTextColor(QColor('#000000'))
                 self.txtStdout.clear()
-                self.txtStdout.setText(QCoreApplication.translate('DialogExportData','Java could not be found. Please <a href="https://java.com/en/download/">install Java</a> and or <a href="#configure">configure a custom java path</a>. We also support the JAVA_HOME environment variable in case you prefer this.'))
+                self.txtStdout.setText(QCoreApplication.translate('DialogImportData','Java could not be found. Please <a href="https://java.com/en/download/">install Java</a> and or <a href="#configure">configure a custom java path</a>. We also support the JAVA_HOME environment variable in case you prefer this.'))
                 self.enable()
                 self.progress_bar.hide()
                 return
@@ -313,11 +329,11 @@ class DialogExportData(QDialog, DIALOG_UI):
             self.buttonBox.setEnabled(True)
             self.buttonBox.addButton(QDialogButtonBox.Close)
             self.progress_bar.setValue(100)
-            self.show_message(QCoreApplication.translate('DialogExportData', 'Export of the data was successfully completed.') , Qgis.Success)
+            self.show_message(QCoreApplication.translate('DialogImportData', 'Import of the data was successfully completed'), Qgis.Success)
 
     def save_configuration(self, configuration):
         settings = QSettings()
-        settings.setValue('Asistente-LADM_COL/QgisModelBaker/ili2pg/xtffile_export', configuration.xtffile)
+        settings.setValue('Asistente-LADM_COL/QgisModelBaker/ili2pg/xtffile_import', configuration.xtffile)
         settings.setValue('Asistente-LADM_COL/QgisModelBaker/importtype', self.type_combo_box.currentData())
         settings.setValue('Asistente-LADM_COL/QgisModelBaker/show_log', 1 if self.log_config.isCollapsed() else 0)
 
@@ -326,16 +342,16 @@ class DialogExportData(QDialog, DIALOG_UI):
 
     def restore_configuration(self):
         settings = QSettings()
-        self.xtf_file_line_edit.setText(settings.value('Asistente-LADM_COL/QgisModelBaker/ili2pg/xtffile_export'))
+        self.xtf_file_line_edit.setText(settings.value('Asistente-LADM_COL/QgisModelBaker/ili2pg/xtffile_import'))
         self.type_combo_box.setCurrentIndex(self.type_combo_box.findData(settings.value('Asistente-LADM_COL/QgisModelBaker/importtype', 'ili2pg')))
         self.type_changed()
 
         # Show log
-        value_show_log = settings.value('Asistente-LADM_COL/QgisModelBaker/show_log') if settings.value('Asistente-LADM_COL/QgisModelBaker/show_log') else 0
+        value_show_log = settings.value('Asistente-LADM_COL/QgisModelBaker/show_log')  if settings.value('Asistente-LADM_COL/QgisModelBaker/show_log') else 0
         self.log_config.setCollapsed(bool(int(value_show_log)))
 
         # set model repository
-        # if there is no option by default use online model repository
+        # if there is no option  by default use online model repository
         custom_model_is_checked =  settings.value('Asistente-LADM_COL/models/custom_model_directories_is_checked') if settings.value('Asistente-LADM_COL/models/custom_model_directories_is_checked') else 0
         self.use_local_models = bool(int(custom_model_is_checked))
         if self.use_local_models:
@@ -346,7 +362,7 @@ class DialogExportData(QDialog, DIALOG_UI):
         Get the configuration that is updated with the user configuration changes on the dialog.
         :return: Configuration
         """
-        configuration = ExportConfiguration()
+        configuration = ImportDataConfiguration()
 
         if self.type_combo_box.currentData() == 'ili2pg':
             # PostgreSQL specific options
@@ -360,6 +376,13 @@ class DialogExportData(QDialog, DIALOG_UI):
             configuration.dbfile = self.gpkg_file_line_edit.text().strip()
 
         configuration.xtffile = self.xtf_file_line_edit.text().strip()
+        configuration.delete_data = False
+
+        configuration.epsg = DEFAULT_EPSG
+        configuration.inheritance = DEFAULT_INHERITANCE
+        configuration.create_basket_col = False
+        configuration.create_import_tid = False
+        configuration.stroke_arcs = True
 
         # Check custom model directories
         if self.use_local_models:
@@ -371,17 +394,11 @@ class DialogExportData(QDialog, DIALOG_UI):
 
         configuration.base_configuration = self.base_configuration
         if self.get_ili_models():
-            configuration.iliexportmodels = ';'.join(self.get_ili_models())
             configuration.ilimodels = ';'.join(self.get_ili_models())
 
         return configuration
 
-    def print_new_info(self, message_error):
-        self.txtStdout.setTextColor(QColor('#000000'))
-        self.txtStdout.clear()
-        self.txtStdout.setText(message_error)
-
-    def print_info(self, text, text_color='#000000'):
+    def print_info(self, text, text_color='#000000', clear=False):
         self.txtStdout.setTextColor(QColor(text_color))
         self.txtStdout.append(text)
         QCoreApplication.processEvents()
@@ -401,12 +418,13 @@ class DialogExportData(QDialog, DIALOG_UI):
     def on_process_finished(self, exit_code, result):
         color = '#004905' if exit_code == 0 else '#aa2222'
         self.txtStdout.setTextColor(QColor(color))
-        self.txtStdout.append(QCoreApplication.translate('DialogExportData','Finished ({})'.format(exit_code)))
-        if result == iliexporter.Exporter.SUCCESS:
+        self.txtStdout.append('Finished ({})'.format(exit_code))
+        if result == iliimporter.Importer.SUCCESS:
             self.buttonBox.clear()
             self.buttonBox.setEnabled(True)
             self.buttonBox.addButton(QDialogButtonBox.Close)
         else:
+            self.show_message(QCoreApplication.translate('DialogImportData', 'Error when importing data'), Qgis.Warning)
             self.enable()
 
     def advance_progress_bar_by_text(self, text):
@@ -418,7 +436,7 @@ class DialogExportData(QDialog, DIALOG_UI):
             QCoreApplication.processEvents()
 
     def show_help(self):
-        self.qgis_utils.show_help("export_data")
+        self.qgis_utils.show_help("import_data")
 
     def disable(self):
         self.type_combo_box.setEnabled(False)
@@ -443,15 +461,3 @@ class DialogExportData(QDialog, DIALOG_UI):
         elif self.type_combo_box.currentData() == 'ili2gpkg':
             self.pg_config.hide()
             self.gpkg_config.show()
-
-    def xtf_browser_opened_to_true(self):
-        """
-        Slot. Sets a flag to true to eventually avoid asking a user whether to overwrite a file.
-        """
-        self.xtf_browser_was_opened = True
-
-    def xtf_browser_opened_to_false(self):
-        """
-        Slot. Sets a flag to false to eventually ask a user whether to overwrite a file.
-        """
-        self.xtf_browser_was_opened = False
