+++ asistente_ladm_col/lib/dbconnector/pg_connector.py
@@ -784,7 +784,7 @@ class PGConnector(DBConnector):
         result = self.execute_sql_query(query)
         return result if not isinstance(result, tuple) else None
 
-    def _get_models(self, dbschema):
-        query = "SELECT modelname FROM {schema}.t_ili2db_model".format(schema=dbschema)
+    def _get_models(self, schema=None):
+        query = "SELECT modelname FROM {schema}.t_ili2db_model".format(schema=schema if schema else self.schema)
         result = self.execute_sql_query(query)
         return result if not isinstance(result, tuple) else None
